enrollment

+++ src/models/Course.java
@@ -1,12 +1,13 @@
 package models;
 
 public class Course {
+
 	private int id;
 	private String courseCode;
-	private int units;
+	private float units;
 	private int maxStudents;
 	private int enrolledStudents;
-	
+
 	public Course(int id, String courseCode, int units, int maxStudents) {
 		super();
 		this.id = id;
@@ -14,8 +15,8 @@ public class Course {
 		this.units = units;
 		this.maxStudents = maxStudents;
 	}
-	
-	public Course(int id, String courseCode, int units, int maxStudents, int enrolledStudents) {
+
+	public Course(int id, String courseCode, float units, int maxStudents, int enrolledStudents) {
 		super();
 		this.id = id;
 		this.courseCode = courseCode;
@@ -23,7 +24,7 @@ public class Course {
 		this.maxStudents = maxStudents;
 		this.enrolledStudents = enrolledStudents;
 	}
-	
+
 	public int getEnrolledStudents() {
 		return enrolledStudents;
 	}
@@ -35,27 +36,33 @@ public class Course {
 	public int getId() {
 		return id;
 	}
+
 	public void setId(int id) {
 		this.id = id;
 	}
+
 	public String getCourseCode() {
 		return courseCode;
 	}
+
 	public void setCourseCode(String courseCode) {
 		this.courseCode = courseCode;
 	}
-	public int getUnits() {
+
+	public float getUnits() {
 		return units;
 	}
+
 	public void setUnits(int units) {
 		this.units = units;
 	}
+
 	public int getMaxStudents() {
 		return maxStudents;
 	}
+
 	public void setMaxStudents(int maxStudents) {
 		this.maxStudents = maxStudents;
 	}
-	
-	
+
 }
